* Co/Fb: Fix bad character cast for characters between 0x80 and 0xFF.
Since the character type is signed, this means that an implicit cast
to an int would extend it to 0xFFFFFF??, rather than 0x000000??. This
is bad.

+++ src/co/fb.c
@@ -36,7 +36,7 @@ void CoVbePlotChar (Console *this, int x, int y, char c)
 	VBEData* backup = g_vbeData;
 	g_vbeData = &g_mainScreenVBEData;
 	
-	VidPlotChar (c, this->offX + (x << 3), this->offY + (y << (3 + (g_uses8by16Font))), g_vgaColorsToRGB[this->color & 0xF], g_vgaColorsToRGB[this->color >> 4]);
+	VidPlotChar ((uint8_t)c, this->offX + (x << 3), this->offY + (y << (3 + (g_uses8by16Font))), g_vgaColorsToRGB[this->color & 0xF], g_vgaColorsToRGB[this->color >> 4]);
 	g_vbeData = backup;
 }
